requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0

# LLM (CPU version - works everywhere)
# For GPU acceleration, run: python scripts/install_llama_gpu.py
llama-cpp-python>=0.2.0

# Retrieval
rank-bm25>=0.2.2

# Data
datasets>=2.14.0
huggingface-hub>=0.17.0

# Evaluation
scikit-learn>=1.3.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.0.0
tqdm>=4.65.0
